+++ package.json
@@ -2,14 +2,20 @@
   "name": "@dataai/msw-auto-mock",
   "version": "0.31.0",
   "description": "Generate random mock data from OpenAPI descriptions for msw.",
-  "main": "dist/index.js",
+  "type": "module",
+  "main": "dist/index.cjs",
   "module": "dist/index.js",
   "types": "dist/index.d.ts",
   "exports": {
     ".": {
       "import": "./dist/index.js",
-      "require": "./dist/index.js",
+      "require": "./dist/index.cjs",
       "types": "./dist/index.d.ts"
+    },
+    "./node": {
+      "import": "./dist/node/node.js",
+      "require": "./dist/node/node.cjs",
+      "types": "./dist/node/node.d.ts"
     }
   },
   "scripts": {
@@ -57,7 +63,7 @@
     "url": "https://github.com/zoubingwu/msw-auto-mock/issues"
   },
   "homepage": "https://github.com/zoubingwu/msw-auto-mock",
-  "files": ["dist"],
+  "files": ["dist", "dist/node"],
   "keywords": ["msw", "mock", "service-worker", "openapi"],
   "publishConfig": {
     "access": "public"
