+++ src/utils.cjs.ts
@@ -1,6 +1,5 @@
 import fs from 'node:fs/promises';
 import path from 'node:path';
-import { camelCase } from 'es-toolkit';
 import prettier from 'prettier';
 
 async function prettify(content: string): Promise<string> {
@@ -20,19 +19,6 @@ async function prettify(content: string): Promise<string> {
     return content;
   }
 }
-
-export const toExpressLikePath = (path: string) =>
-  // use `.+?` for lazy match
-  path.replace(/{(.+?)}/g, (_match, p1: string) => `:${camelCase(p1)}`);
-
-export const isValidRegExp = (pattern: string): boolean => {
-  try {
-    new RegExp(pattern);
-    return true;
-  } catch {
-    return false;
-  }
-};
 
 export const writeFile = async (filePath: string, content: string) => {
   const prettifiedContent = await prettify(content);
